Further refactor of photo resubmission components
