+++ src/Patient/ReportingFlows/AltPhotoFlows/PreSubmissionView.js
@@ -1,0 +1,96 @@
+import React, { useState } from 'react';
+import TestStripPhotoInfo from '../../../Components/Patient/TestStripPhotoInfo';
+import PhotoPrompt from '../../../Components/Patient/PhotoPrompt';
+import { makeStyles } from '@material-ui/core/styles';
+import Camera from '../../../ImageCapture/Camera';
+import ClickableText from '../../../Basics/ClickableText';
+import ReplayIcon from '@material-ui/icons/Replay';
+import { useTranslation } from 'react-i18next';
+import BottomButton from './BottomButton';
+import Box from '@material-ui/core/Box';
+import Grid from '@material-ui/core/Grid';
+import PermissionsError from '../../../ImageCapture/PermissionsError';
+
+import BackSubmissionText from './BackSubmissionText';
+import NotEligible from './NotEligible';
+import { CircularProgress } from '@material-ui/core';
+
+const useStyles = makeStyles({
+    strip: {
+        display: "block",
+        height: '50vh',
+        width: '90%',
+        "& > img": {
+            objectFit: 'contain',
+            height: '100%',
+            width: '100%'
+        },
+        margin: 'auto'
+    },
+    fullWidth: {
+        padding: "0 1em",
+        boxSizing: "border-box"
+    }
+})
+
+const PreSubmissionView = ({ photo, eligible, setPhoto, handleSubmit, requestDateFormatted, loading }) => {
+
+    const { t } = useTranslation('translation');
+    const classes = useStyles();
+    const [cameraOpen, setCameraOpen] = useState(false);
+    const [permissionsError, setPermissionsError] = useState(false);
+
+    const handleExit = () => {
+        setCameraOpen(false);
+    }
+
+    const handlePhoto = (newPhoto) => {
+        setPhoto(newPhoto);
+        setCameraOpen(false);
+    }
+
+    const handlePermissionsError = () => {
+        handleExit();
+        setPermissionsError(true);
+    }
+
+    if (loading) {
+        return (<Loading />);
+    }
+
+    return (<>
+        {photo ? <>
+            <div className={classes.strip}>
+                <img src={photo} />
+            </div>
+            <div className={classes.fullWidth}>
+                <ClickableText onClick={() => { setCameraOpen(true) }} icon={<ReplayIcon />} text={t('patient.report.photo.retakePhoto')} />
+            </div>
+            <BottomButton disabled={(eligible && !photo)} onClick={handleSubmit}>{t('coordinator.patientProfile.editDetails.submit')}</BottomButton>
+        </> : <>
+            {eligible ? <>
+                <BackSubmissionText photo={photo !== false} requestDateFormatted={requestDateFormatted} />
+                <PhotoPrompt onClick={() => { setCameraOpen(true) }} />
+                {permissionsError && <PermissionsError />}
+                <Box height=".5em" />
+                <TestStripPhotoInfo showSkipOptions={false} />
+            </> : <NotEligible />}
+        </>}
+        {cameraOpen && <Camera
+            handlePermissionsError={handlePermissionsError}
+            handleExit={handleExit}
+            returnPhoto={handlePhoto} />}
+    </>)
+}
+
+
+const Loading = () => {
+    const classes = useStyles();
+    return (
+        <Grid className={classes.loading} alignItems="center" justify="center" container>
+            <CircularProgress variant="indeterminate" />
+        </Grid>
+    )
+}
+
+export default PreSubmissionView;
